requirements.txt
absl-py==0.14.1
aiohttp==3.9.3
aiosignal==1.3.1
altair==4.0.0
annotated-types==0.6.0
anyascii==0.3.2
anyio==4.2.0
appdirs==1.4.4
attrs==23.2.0
audioread==3.0.1
Babel==2.14.0
bangla==0.0.2
bleach==6.1.0
blinker==1.7.0
blis==0.7.11
bnnumerizer==0.0.2
bnunicodenormalizer==0.1.6
cachetools==5.3.2
catalogue==2.0.10
certifi==2022.12.7
cffi==1.16.0
charset-normalizer==2.1.1
click==8.1.3
cloudpathlib==0.16.0
colorama==0.4.6
confection==0.1.4
contourpy==1.1.0
coqpit==0.0.17

dataclasses-json==0.5.14
dateparser==1.1.8
decorator==5.1.1
docopt==0.6.2
einops==0.4.1
encodec==0.1.1
entrypoints==0.4
faiss-cpu==1.7.4
fastjsonschema==2.19.1
ffmpeg==1.4
filelock==3.9.0
Flask==2.2.3
fonttools==4.41.1
frozenlist==1.4.1
fsspec==2023.12.2
g2pkk==0.1.2
gitdb==4.0.11
GitPython==3.1.41
google-auth==2.27.0
google-auth-oauthlib==1.2.0
greenlet==3.0.3
grpcio==1.60.0
gruut==2.3.4
gruut-ipa==0.13.0
gruut_lang_de==2.0.0
gruut_lang_en==2.0.0
gruut_lang_es==2.0.0
gruut_lang_fr==2.0.2
hangul-romanize==0.1.0
hjson==3.1.0
huggingface-hub==0.14.1
idna==3.4
importlib-metadata==7.0.1
inflect==7.0.0
InstructorEmbedding==1.0.1
itsdangerous==2.1.2
jamo==0.4.1
jieba==0.42.1
Jinja2==3.1.2
joblib==1.3.2
jsonlines==1.2.0
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_core==5.7.1
kiwisolver==1.4.4
langchain==0.1.7
langchain-community==0.0.20
langchain-core==0.1.23
langcodes==3.3.0
langsmith==0.0.87
lazy_loader==0.3
librosa==0.9.1
llvmlite==0.41.1
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.2
marshmallow==3.20.2
matplotlib==3.8.2
mdurl==0.1.2
mistune==3.0.2
mpmath==1.3.0
msgpack==1.0.7
multidict==6.0.4
murmurhash==1.0.10
mutagen==1.47.0
mypy-extensions==1.0.0
nbconvert==5.3.1
nbformat==5.9.2
networkx==2.8.8
nltk==3.8.1
num2words==0.5.13
numba==0.58.1
numexpr==2.9.0
numpy==1.26.0
oauthlib==3.2.2
openai==0.27.6
openapi-schema-pydantic==1.2.4
packaging==23.2
pandas==1.5.3
pandocfilters==1.5.1
pdfminer.six==20221105
pdfplumber==0.10.3
pillow==10.2.0
platformdirs==4.1.0
pooch==1.8.0
preshed==3.0.9
progressbar==2.5
protobuf==3.20.3
psutil==5.9.8
py-cpuinfo==9.0.0
pyarrow==15.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0

pycparser==2.21
pydantic==1.10.14
pydantic_core==2.14.6
pydeck==0.8.1b0
Pygments==2.17.2
Pympler==1.0.1
PyMuPDF==1.23.21
PyMuPDFb==1.23.9
pynndescent==0.5.11
pyparsing==3.0.9
PyPDF2==2.1.1
pypdfium2==4.27.0
pypinyin==0.50.0
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
pysbd==0.3.4
python-crfsuite==0.9.10
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.4
huggingface
PyYAML==6.0.0
pyzmq==25.1.2
referencing==0.33.0
regex==2023.12.25
requests==2.28.1
requests-oauthlib==1.3.1
resampy==0.4.2
rich==13.7.0
streamlit==1.32.2
sentence-transformers==2.0.0

